[코드 추가] 버튼 오버 및 준비 완료 연동 구현, 대기실에서 체력바가 출력되는 버그 픽스(with 김민규, 전성희)

+++ TermProject/Button.cpp
@@ -30,7 +30,7 @@ void CButton::Render(HDC hMemDC, HDC hMemDC2)
 		break;
 	}
 
-	DrawRect(hMemDC, GetPosition(), BitmapRect.m_Width, BitmapRect.m_Height, hMemDC2, BitmapRect, CFileManager::GetInstance()->GetTransparentColor());
+	DrawRect(hMemDC, GetPosition(), (float)BitmapRect.m_Width, (float)BitmapRect.m_Height, hMemDC2, BitmapRect, CFileManager::GetInstance()->GetTransparentColor());
 }
 
 void CButton::SetActive(bool IsActive)
